pyproject.toml
[project]
name = "orbit-assist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"fastapi>=0.115.0",
	"google-genai>=1.67.0",
	"httpx>=0.28.1",
	"pydantic-settings>=2.11.0",
	"psycopg[binary]>=3.2.10",
	"psycopg-pool>=3.2.7",
	"uvicorn[standard]>=0.30.0",
]

[project.scripts]
dev = "orbit_assist.dev:run"


[tool.uv]
package = true

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["orbit_assist*"]